get_auth_data_method implementation

+++ crates/auth/auth_domain/src/model/auth_data.rs
@@ -1,6 +1,6 @@
 #[derive(PartialEq, Eq, Debug)]
 pub struct AuthData {
-    pub id: i128,
+    pub id: i64,
     pub name: String,
     pub email: String,
     pub password: String,
